add avatar,  mvc and improvement
