+++ templates/profile.html
@@ -5,74 +5,91 @@
 {% endblock %}
 
 {% block main %}
-<div class="separator-nav">
-</div>
-<div class="thank-main">
-  <div id="profile-box">
-    <div class="decorator-bar">
-    </div>
-    <div class="dialog-main-container">
-      <div class="dialog-main">
-        <div class="dialog-title">會員資料</div>
-        <div class="profile-label">姓名</div>
-        <div class="input"><input type="text" class="input-text" id="profile-name" name="profile-name" required>
+<div class="content">
+  <div class="separator-nav">
+  </div>
+  <div class="thank-main">
+    <div id="profile-box">
+      <div class="decorator-bar">
+      </div>
+      <div class="dialog-main-container">
+        <div class="dialog-main">
+          <div class="avatar-container">
+            <img class="avatar" id="avatar-upload" src="/uploads/default.png">
+          </div>
+          <div class="edit-avatar"><input class="upload-file" type="file" name="image" accept="image/*"
+              onchange="readURL(this)"></div>
+          <div class="profile-label">姓名
+          </div>
+          <div class="input"><input type="text" class="input-text" id="profile-name" name="profile-name" required>
+          </div>
+          <div class="profile-label">email</div>
+          <div class="input-email"><input type="text" class="input-text" id="profile-email" name="profile-email"
+              required>
+          </div>
+          <div class="profile-label">密碼</div>
+          <div class="input"><input type="password" class="input-text" id="profile-password" name="profile-password"
+              placeholder="請輸入想更改的密碼">
+          </div>
+          <button class="dialog-button" onclick="update()">修改會員資料</button>
+          <div class="message" id="profile-message"></div>
         </div>
-        <div class="profile-label">email</div>
-        <div class="input-email"><input type="text" class="input-text" id="profile-email" name="profile-email" required>
-        </div>
-        <div class="profile-label">密碼</div>
-        <div class="input"><input type="password" class="input-text" id="profile-password" name="profile-password"
-            placeholder="請輸入想更改的密碼">
-        </div>
-        <button class="dialog-button" onclick="update()">修改會員資料</button>
-        <div class="message" id="profile-message"></div>
       </div>
     </div>
   </div>
 </div>
+
 {% endblock %}
 
 
 {% block js %}
 <script>
 
+  function readURL(input) {
+    if (input.files && input.files[0]) {
+      let reader = new FileReader();
+      reader.onload = function (e) {
+        let img = document.querySelector("#avatar-upload");
+        img.setAttribute("src", e.target.result)
+      }
+      reader.readAsDataURL(input.files[0]);
+    }
+  }
+
   function update() {
-    getStatus()
+    longinStatus
       .then(userInfo => {
 
         if (userInfo === null) {
           window.location.href = '/'
         };
 
-        let body = JSON.stringify({
-          name: document.querySelector("#profile-name").value,
-          email: document.querySelector("#profile-email").value,
-        })
+        const formData = new FormData();
+        formData.append('file', document.querySelector('.upload-file').files[0]);
+        formData.append('name', document.querySelector("#profile-name").value);
+        formData.append('email', document.querySelector("#profile-email").value);
 
         if (document.querySelector("#profile-password").value.trim() !== "") {
-          body = JSON.stringify({
-            name: document.querySelector("#profile-name").value,
-            email: document.querySelector("#profile-email").value,
-            password: document.querySelector("#profile-password").value
-          })
+          formData.append('password', document.querySelector("#profile-password").value);
         }
 
         const putMethod = {
           method: 'PUT',
           headers: {
-            'Content-Type': 'application/json',
+            // 'Content-Type': 'application/json',
             'Authorization': 'Bearer ' + localStorage.getItem('token')
           },
-          body: body
+          body: formData,
         }
 
-        fetch('/api/profile', putMethod)
+        fetch('/api/user/profile', putMethod)
           .then(response => response.json())
           .then(responseJson => {
             if (responseJson["ok"]) {
+              getStatus();
               document.querySelector("#profile-message").textContent = "會員資料已修改成功";
               document.querySelector("#profile-message").style.display = 'flex';
-              document.querySelector("#profile-box").style.height = '430px';
+              document.querySelector("#profile-box").style.height = '577px';
             } else {
               alert(responseJson["message"]);
             }
@@ -85,20 +102,24 @@
   }
 
   function getProfile() {
-    getStatus()
+    longinStatus
       .then(userInfo => {
 
         if (userInfo === null) {
           window.location.href = '/'
         };
 
+        document.querySelector(".content").style.display = "block";
+        if (userInfo["filename"] !== null) {
+          document.querySelector("#avatar-upload").src = "/uploads/" + userInfo["id"] + "/" + userInfo["filename"];
+        }
         document.querySelector("#profile-name").value = userInfo["name"];
         document.querySelector("#profile-email").value = userInfo["email"];
         document.querySelector("#profile-name").style.color = "black";
         document.querySelector("#profile-email").style.color = "black";
         document.querySelector("#profile-password").style.color = "black";
         document.querySelector("#profile-message").style.display = 'none';
-
+        document.querySelector(".attraction-footer").style.height = "100%";
       })
       .catch(error => console.error("Error:", error))
   }
