+++ templates/thankyou.html
@@ -11,7 +11,7 @@
   <div class="thank-container">
     <div class="main-words" id="msg"></div>
     <div class="main-words" id="orderId"></div>
-    <div class="sub-words">※可至會員中心，以「訂單編號」查詢訂單詳情</div>
+    <div class="sub-words">※訂單詳情：可至會員中心以「訂單編號」查詢</div>
   </div>
 </div>
 
@@ -24,14 +24,14 @@
 
   function paymentResult() {
 
-    getStatus()
+    longinStatus
       .then(userInfo => {
 
         if (userInfo === null) {
           window.location.href = '/'
         };
 
-        document.querySelector(".attraction-footer").style.height = "100%";
+
         const words = decodeURIComponent(window.location.search).split("&");
         const orderId = words[0].slice(words[0].indexOf("=") + 1);
         const msg = words[1].slice(words[1].indexOf("=") + 1);
